kn: Move contents of zxingUitls.kt and ReaderOptions.kt to BarcodeReader.kt
